Add files via upload
Developed a Text to Speech Converter using HTML, CSS, JavaScript, and Fetch API. The app allows users to input text, which is then
converted into audible speech. This project provided a hands-on experience in web development, demonstrating my ability to integrate
APIs and create accessible web applications that cater to a wide range of user needs.